[FIX] 홈화면 썸네일 뷰 → 텍스트에 **로 볼드 표시 (#208)
